+++ swak/text/misc.py
@@ -1,4 +1,9 @@
+from typing import Any, Literal
 from enum import StrEnum
+
+type Toml = dict[str, Any]
+type Yaml = dict[str, Any] | list[Any]
+type NotFoundLiteral = Literal[NotFound.IGNORE, NotFound.WARN, NotFound.RAISE]
 
 
 class NotFound(StrEnum):
@@ -6,7 +11,8 @@ class NotFound(StrEnum):
 
     See Also
     --------
-    TomlReader, YamlReader
+    TomlReader
+    YamlReader
 
     """
     IGNORE = 'ignore'